fixed clippy errors

+++ src/codes/zeta.rs
@@ -30,11 +30,19 @@ pub fn len_zeta<const USE_TABLE: bool>(mut value: u64, k: u64) -> usize {
     len_unary::<false>(h) + len_minimal_binary(value - l, u - l)
 }
 
-/// 
+/// Trait for objects that can read Zeta codes
 pub trait ZetaRead<BO: BitOrder>: MinimalBinaryRead<BO> {
     /// Generic ζ code reader
+    /// 
+    /// # Errors
+    /// This function fails only if the BitRead backend has problems reading
+    /// bits, as when the stream ended unexpectedly
     fn read_zeta<const USE_TABLE: bool>(&mut self, k: u64) -> Result<u64>;
     /// Specialized ζ code reader for k = 3
+    /// 
+    /// # Errors
+    /// This function fails only if the BitRead backend has problems reading
+    /// bits, as when the stream ended unexpectedly
     fn read_zeta3<const USE_TABLE: bool>(&mut self) -> Result<u64>;
 }
 
@@ -50,7 +58,7 @@ impl<B: BitRead<M2L>> ZetaRead<M2L> for B {
             if let Ok(idx) = self.peek_bits(zeta_tables::READ_BITS) {
                 let (value, len) = zeta_tables::READ_M2L[idx as usize];
                 if len != zeta_tables::MISSING_VALUE_LEN {
-                    self.skip_bits(len as u8)?;
+                    self.skip_bits(len)?;
                     return Ok(value as u64);
                 }
             }
@@ -70,7 +78,7 @@ impl<B: BitRead<L2M>> ZetaRead<L2M> for B {
             if let Ok(idx) = self.peek_bits(zeta_tables::READ_BITS) {
                 let (value, len) = zeta_tables::READ_L2M[idx as usize];
                 if len != zeta_tables::MISSING_VALUE_LEN {
-                    self.skip_bits(len as u8)?;
+                    self.skip_bits(len)?;
                     return Ok(value as u64);
                 }
             }
@@ -81,7 +89,6 @@ impl<B: BitRead<L2M>> ZetaRead<L2M> for B {
 
 #[inline(always)]
 fn default_read_zeta<BO: BitOrder, B: BitRead<BO>>(backend: &mut B, k: u64) -> Result<u64> {
-    // implementation taken from github.com/vigna/dsiutils @ InputBitStram.java
     let h = backend.read_unary::<false>()?;
     let u = fast_pow_2((h + 1) * k);
     let l = fast_pow_2(h * k);
@@ -89,11 +96,19 @@ fn default_read_zeta<BO: BitOrder, B: BitRead<BO>>(backend: &mut B, k: u64) -> R
     Ok(l + res - 1)
 }
 
-/// 
+/// Trait for objects that can write Zeta codes
 pub trait ZetaWrite<BO: BitOrder>: MinimalBinaryWrite<BO> {
     /// Generic ζ code writer
+    /// 
+    /// # Errors
+    /// This function fails only if the BitWrite backend has problems writing
+    /// bits, as when the stream ended unexpectedly
     fn write_zeta<const USE_TABLE: bool>(&mut self, value: u64, k: u64) -> Result<()>;
     /// Specialized ζ code writer for k = 3
+    /// 
+    /// # Errors
+    /// This function fails only if the BitWrite backend has problems writing
+    /// bits, as when the stream ended unexpectedly
     fn write_zeta3<const USE_TABLE: bool>(&mut self, value: u64) -> Result<()>;
 }
